% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concaveman-pkg.R, R/concaveman.R
\docType{package}
\name{concaveman3d}
\alias{concaveman3d-package}
\alias{concaveman3d}
\alias{concaveman3d.matrix}
\alias{concaveman3d.data.frame}
\title{concaveman: A very fast 3D concave hull algorithm.}
\usage{
concaveman3d(points, concavity, length_threshold)

\method{concaveman3d}{matrix}(points, concavity = 2, length_threshold = 0)

\method{concaveman3d}{data.frame}(points, concavity = 2, length_threshold = 0)
}
\arguments{
\item{points}{the points for which the concave hull must be computed. Can be represented as a matrix of coordinates or an \code{sf} object.}

\item{concavity}{a relative measure of concavity. 1 results in a relatively detailed shape, Infinity results in a convex hull. You can use values lower than 1, but they can produce pretty crazy shapes.}

\item{length_threshold}{when a segment length is under this threshold, it stops being considered for further detalization. Higher values result in simpler shapes.}
}
\value{
an object of the same class as \code{points}: a matrix of coordinates or a data.frame.
}
\description{
This package is a simple R port (through \href{https://github.com/KayJay7/concaveman3d}{\code{V8}} of a \href{https://github.com/mapbox/concaveman}{JavaScript library by Vladimir Agafonkin}.

3D implementation of the Park-Oh algorithm (https://jise.iis.sinica.edu.tw/JISESearch/fullText?pId=245&code=5A9B97538372AA1)
}
\examples{
data(points)
polygons <- concaveman(points)

}
\seealso{
Useful links:
\itemize{
  \item \url{https://joelgombin.github.io/concaveman/}
  \item \url{http://www.github.com/joelgombin/concaveman/}
  \item Report bugs at \url{http://www.github.com/joelgombin/concaveman/issues}
}

}
\author{
\strong{Maintainer}: Joël Gombin \email{joel.gombin@gmail.com}

Authors:
\itemize{
  \item Ramnath Vaidyanathan
  \item Vladimir Agafonkin
}

Other contributors:
\itemize{
  \item Mapbox [copyright holder]
}

}
